maintenance model part

+++ database/migrations/2024_03_05_231619_create_department_course_parts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->text('note')->nullable();
             $table->integer('score')->nullable();
             $table->integer('lectures_count')->nullable();
-            $table->bigInteger('lecture_duration')->nullable();
+            $table->integer('lecture_duration')->nullable();
 
             // Define composite foreign key constraint
             $table->unsignedBigInteger('course_part_id');
